mistakes fixed
